requirements.txt
PyYAML==6.0.2
requests==2.32.3
python-dotenv==1.0.1
